+++ src/cmdliner_base.ml
@@ -146,61 +146,68 @@ let err_sep_miss sep s =
 
 type 'a parser = string -> [ `Ok of 'a | `Error of string ]
 type 'a printer = Format.formatter -> 'a -> unit
-type 'a conv = 'a parser * 'a printer
+type 'a conv = {
+  parse: 'a parser;
+  print: 'a printer;
+}
 
-let some ?(none = "") (parse, print) =
-  let parse s = match parse s with `Ok v -> `Ok (Some v) | `Error _ as e -> e in
+let some ?(none = "") conv =
+  let parse s = match conv.parse s with `Ok v -> `Ok (Some v) | `Error _ as e -> e in
   let print ppf v = match v with
   | None -> Format.pp_print_string ppf none
-  | Some v -> print ppf v
+  | Some v -> conv.print ppf v
   in
-  parse, print
+  {parse; print}
 
-let some' ?none (parse, print) =
-  let parse s = match parse s with `Ok v -> `Ok (Some v) | `Error _ as e -> e in
+let some' ?none conv =
+  let parse s = match conv.parse s with `Ok v -> `Ok (Some v) | `Error _ as e -> e in
   let print ppf = function
-  | None -> (match none with None -> () | Some v -> print ppf v)
-  | Some v -> print ppf v
+  | None -> (match none with None -> () | Some v -> conv.print ppf v)
+  | Some v -> conv.print ppf v
   in
-  parse, print
+  {parse; print}
 
 let bool =
   let parse s = try `Ok (bool_of_string s) with
   | Invalid_argument _ ->
       `Error (err_invalid_val s (alts_str ~quoted:true ["true"; "false"]))
   in
-  parse, Format.pp_print_bool
+  {parse; print=Format.pp_print_bool}
 
 let char =
   let parse s = match String.length s = 1 with
   | true -> `Ok s.[0]
   | false -> `Error (err_invalid_val s "expected a character")
   in
-  parse, pp_char
+  {parse; print=pp_char}
 
 let parse_with t_of_str exp s =
   try `Ok (t_of_str s) with Failure _ -> `Error (err_invalid_val s exp)
 
 let int =
-  parse_with int_of_string "expected an integer", Format.pp_print_int
+  let parse = parse_with int_of_string "expected an integer" in
+  {parse; print=Format.pp_print_int}
 
 let int32 =
-  parse_with Int32.of_string "expected a 32-bit integer",
-  (fun ppf -> pp ppf "%ld")
+  let parse = parse_with Int32.of_string "expected a 32-bit integer" in
+  let print ppf = pp ppf "%ld" in
+  {parse; print}
 
 let int64 =
-  parse_with Int64.of_string "expected a 64-bit integer",
-  (fun ppf -> pp ppf "%Ld")
+  let parse = parse_with Int64.of_string "expected a 64-bit integer" in
+  let print ppf = pp ppf "%Ld" in
+  {parse; print}
 
 let nativeint =
-  parse_with Nativeint.of_string "expected a processor-native integer",
-  (fun ppf -> pp ppf "%nd")
+  let parse = parse_with Nativeint.of_string "expected a processor-native integer" in
+  let print ppf = pp ppf "%nd" in
+  {parse; print}
 
 let float =
-  parse_with float_of_string "expected a floating point number",
-  Format.pp_print_float
+  let parse = parse_with float_of_string "expected a floating point number" in
+  {parse; print=Format.pp_print_float}
 
-let string = (fun s -> `Ok s), pp_str
+let string = {parse=(fun s -> `Ok s); print=pp_str}
 let enum sl =
   if sl = [] then invalid_arg err_empty_list else
   let t = Cmdliner_trie.of_list sl in
@@ -218,28 +225,28 @@ let enum sl =
     try pp_str ppf (List.assoc v sl_inv)
     with Not_found -> invalid_arg (err_incomplete_enum (List.map fst sl))
   in
-  parse, print
+  {parse; print}
 
 let file =
   let parse s = match Sys.file_exists s with
   | true -> `Ok s
   | false -> `Error (err_no "file or directory" s)
   in
-  parse, pp_str
+  {parse; print=pp_str}
 
 let dir =
   let parse s = match Sys.file_exists s with
   | true -> if Sys.is_directory s then `Ok s else `Error (err_not_dir s)
   | false -> `Error (err_no "directory" s)
   in
-  parse, pp_str
+  {parse; print=pp_str}
 
 let non_dir_file =
   let parse s = match Sys.file_exists s with
   | true -> if not (Sys.is_directory s) then `Ok s else `Error (err_is_dir s)
   | false -> `Error (err_no "file" s)
   in
-  parse, pp_str
+  {parse; print=pp_str}
 
 let split_and_parse sep parse s = (* raises [Failure] *)
   let parse sub = match parse sub with
@@ -257,25 +264,25 @@ let split_and_parse sep parse s = (* raises [Failure] *)
   in
   split [] (String.length s - 1)
 
-let list ?(sep = ',') (parse, pp_e) =
-  let parse s = try `Ok (split_and_parse sep parse s) with
+let list ?(sep = ',') conv =
+  let parse s = try `Ok (split_and_parse sep conv.parse s) with
   | Failure e -> `Error (err_element "list" s e)
   in
   let rec print ppf = function
-  | v :: l -> pp_e ppf v; if (l <> []) then (pp_char ppf sep; print ppf l)
+  | v :: l -> conv.print ppf v; if (l <> []) then (pp_char ppf sep; print ppf l)
   | [] -> ()
   in
-  parse, print
+  {parse; print}
 
-let array ?(sep = ',') (parse, pp_e) =
-  let parse s = try `Ok (Array.of_list (split_and_parse sep parse s)) with
+let array ?(sep = ',') conv =
+  let parse s = try `Ok (Array.of_list (split_and_parse sep conv.parse s)) with
   | Failure e -> `Error (err_element "array" s e)
   in
   let print ppf v =
     let max = Array.length v - 1 in
-    for i = 0 to max do pp_e ppf v.(i); if i <> max then pp_char ppf sep done
+    for i = 0 to max do conv.print ppf v.(i); if i <> max then pp_char ppf sep done
   in
-  parse, print
+  {parse; print}
 
 let split_left sep s =
   try
@@ -284,36 +291,36 @@ let split_left sep s =
     Some ((String.sub s 0 i), (String.sub s (i + 1) (len - i - 1)))
   with Not_found -> None
 
-let pair ?(sep = ',') (pa0, pr0) (pa1, pr1) =
-  let parser s = match split_left sep s with
+let pair ?(sep = ',') conv0 conv1 =
+  let parse s = match split_left sep s with
   | None -> `Error (err_sep_miss sep s)
   | Some (v0, v1) ->
-      match pa0 v0, pa1 v1 with
+      match conv0.parse v0, conv1.parse v1 with
       | `Ok v0, `Ok v1 -> `Ok (v0, v1)
       | `Error e, _ | _, `Error e -> `Error (err_element "pair" s e)
   in
-  let printer ppf (v0, v1) = pp ppf "%a%c%a" pr0 v0 sep pr1 v1 in
-  parser, printer
+  let print ppf (v0, v1) = pp ppf "%a%c%a" conv0.print v0 sep conv1.print v1 in
+  {parse; print}
 
 let t2 = pair
-let t3 ?(sep = ',') (pa0, pr0) (pa1, pr1) (pa2, pr2) =
+let t3 ?(sep = ',') conv0 conv1 conv2 =
   let parse s = match split_left sep s with
   | None -> `Error (err_sep_miss sep s)
   | Some (v0, s) ->
       match split_left sep s with
       | None -> `Error (err_sep_miss sep s)
       | Some (v1, v2) ->
-          match pa0 v0, pa1 v1, pa2 v2 with
+          match conv0.parse v0, conv1.parse v1, conv2.parse v2 with
           | `Ok v0, `Ok v1, `Ok v2 -> `Ok (v0, v1, v2)
           | `Error e, _, _ | _, `Error e, _ | _, _, `Error e ->
               `Error (err_element "triple" s e)
   in
   let print ppf (v0, v1, v2) =
-    pp ppf "%a%c%a%c%a" pr0 v0 sep pr1 v1 sep pr2 v2
+    pp ppf "%a%c%a%c%a" conv0.print v0 sep conv1.print v1 sep conv2.print v2
   in
-  parse, print
+  {parse; print}
 
-let t4 ?(sep = ',') (pa0, pr0) (pa1, pr1) (pa2, pr2) (pa3, pr3) =
+let t4 ?(sep = ',') conv0 conv1 conv2 conv3 =
   let parse s = match split_left sep s with
   | None -> `Error (err_sep_miss sep s)
   | Some(v0, s) ->
@@ -323,15 +330,15 @@ let t4 ?(sep = ',') (pa0, pr0) (pa1, pr1) (pa2, pr2) (pa3, pr3) =
           match split_left sep s with
           | None -> `Error (err_sep_miss sep s)
           | Some (v2, v3) ->
-              match pa0 v0, pa1 v1, pa2 v2, pa3 v3 with
+              match conv0.parse v0, conv1.parse v1, conv2.parse v2, conv3.parse v3 with
               | `Ok v1, `Ok v2, `Ok v3, `Ok v4 -> `Ok (v1, v2, v3, v4)
               | `Error e, _, _, _ | _, `Error e, _, _ | _, _, `Error e, _
               | _, _, _, `Error e -> `Error (err_element "quadruple" s e)
   in
   let print ppf (v0, v1, v2, v3) =
-    pp ppf "%a%c%a%c%a%c%a" pr0 v0 sep pr1 v1 sep pr2 v2 sep pr3 v3
+    pp ppf "%a%c%a%c%a%c%a" conv0.print v0 sep conv1.print v1 sep conv2.print v2 sep conv3.print v3
   in
-  parse, print
+  {parse; print}
 
 let env_bool_parse s = match String.lowercase_ascii s with
 | "" | "false" | "no" | "n" | "0" -> `Ok false